move map/set/multimap from util into collections
