+++ lalrpop/src/collections/mod.rs
@@ -1,0 +1,7 @@
+mod map;
+mod multimap;
+mod set;
+
+pub use self::map::{map, Map};
+pub use self::multimap::Multimap;
+pub use self::set::{set, Set};
