+++ lalrpop/src/util.rs
@@ -1,6 +1,4 @@
-use std::collections::{btree_map, BTreeMap, BTreeSet};
 use std::fmt::{Display, Formatter, Error};
-use std::iter::FromIterator;
 
 pub use std::collections::btree_map as map;
 pub struct Sep<S>(pub &'static str, pub S);
@@ -46,65 +44,6 @@ impl<'a,S:Display> Display for Prefix<&'a [S]> {
         }
         Ok(())
     }
-}
-
-pub struct Multimap<K,V> {
-    map: Map<K,Vec<V>>
-}
-
-impl<K:Ord,V> Multimap<K,V> {
-    pub fn new() -> Multimap<K,V> {
-        Multimap { map: map() }
-    }
-
-    pub fn is_empty(&self) -> bool {
-        self.map.is_empty()
-    }
-
-    pub fn push(&mut self, key: K, value: V) {
-        self.map.entry(key).or_insert(vec![]).push(value);
-    }
-
-    pub fn into_iter(self) -> btree_map::IntoIter<K, Vec<V>> {
-        self.map.into_iter()
-    }
-}
-
-impl<K:Ord,V> IntoIterator for Multimap<K,V> {
-    type Item = (K, Vec<V>);
-    type IntoIter = btree_map::IntoIter<K, Vec<V>>;
-    fn into_iter(self) -> btree_map::IntoIter<K, Vec<V>> {
-        self.into_iter()
-    }
-}
-
-impl<K:Ord,V> FromIterator<(K,V)> for Multimap<K,V> {
-    fn from_iter<T>(iterator: T) -> Self where T: IntoIterator<Item=(K,V)> {
-        let mut map = Multimap::new();
-        for (key, value) in iterator {
-            map.push(key, value);
-        }
-        map
-    }
-}
-
-/// In general, we avoid coding directly against any particular map,
-/// but rather build against `util::Map` (and `util::map` to construct
-/// an instance). This should be a deterministic map, such that two
-/// runs of LALRPOP produce the same output, but otherwise it doesn't
-/// matter much. I'd probably prefer to use `HashMap` with an
-/// alternative hasher, but that's not stable.
-pub type Map<K,V> = BTreeMap<K,V>;
-
-pub fn map<K:Ord,V>() -> Map<K,V> {
-    Map::<K,V>::default()
-}
-
-/// As `Map`, but for sets.
-pub type Set<K> = BTreeSet<K>;
-
-pub fn set<K:Ord>() -> Set<K> {
-    Set::<K>::default()
 }
 
 /// Strip leading and trailing whitespace.
